+++ cmd/operator-sdk/alpha/cmd.go
@@ -15,9 +15,11 @@
 package alpha
 
 import (
-	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/bundle"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/kubebuilder"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/olm"
+	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/bundle"
+	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/cleanup"
+	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/run"
 
 	"github.com/spf13/cobra"
 )
@@ -31,6 +33,8 @@ func NewCmd() *cobra.Command {
 	cmd.AddCommand(
 		olm.NewCmd(),
 		kubebuilder.NewCmd(),
+		run.NewCmd(),
+		cleanup.NewCmd(),
 		bundle.NewCmd(),
 	)
 	return cmd